% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.stat.desc.R
\name{add.stat}
\alias{add.stat}
\alias{add.stat.desc}
\title{add.stat method for desc object}
\usage{
add.stat(tab, data, func.stat, func.stat.name, ...)

\method{add.stat}{desc}(tab, data, func.stat, func.stat.name, ...)
}
\arguments{
\item{tab}{A desc object}

\item{data}{The data frame used to creat tab}

\item{func.stat}{Passed to \code{report.quanti} function}

\item{func.stat.name}{Passed to \code{report.quanti} function}

\item{...}{Other parameters}
}
\value{
A "quanti" desc object (desc$type.desc="quanti")
}
\description{
\code{add.stat.desc} 
Enables the user to add a specific statistic in a desc object.
}
\details{
Enables the user to add a specific statistic in a desc object (only works for quantitative statistics for now).
It calls the report.quanti function with the same attributes as the tab argument and use the reproup function.
to bind the two results.
}
\examples{

# Load the data

data(data)

# The default statistics are given here:

tab1=report.quanti(data=data,y="y_numeric",x1="GROUP",total=TRUE,subjid="SUBJID")

# Define the function corresponding to the coefficient of variation for example

cv=function(y) sd(y,na.rm=TRUE)/mean(y,na.rm=TRUE)

# We use the add.stat function:

tab1.cv=add.stat(tab1,data,func.stat=cv,func.stat.name="Coef. Var")

tab1.cv

# Same with 2 explicative variables

tab=report.quanti(data=data,y="y_numeric",x1="GROUP",
x2="TIMEPOINT",total=TRUE,subjid="SUBJID",
	at.row="TIMEPOINT")

add.stat(tab,data,func.stat=cv,func.stat.name="Coef. Var")


}
\seealso{
\code{\link{regroup}} \code{\link{report.quanti}}
}
