% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adverse_event}
\alias{adverse_event}
\title{adverse_event}
\format{A data frame
#' \describe{
\item{SOCNAME}{Fake System Organ Class variable}
\item{HLTNAME}{Fake factor}
\item{PTNAME}{Fake Prefered Terms variable}
\item{randtrt}{Fake treatment group factor}
\item{SUBJID}{Fake subject Id}
}}
\usage{
data(adverse_event)
}
\description{
Fake adverse event data example
}
\keyword{datasets}
