% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.format.R
\name{apply.format}
\alias{apply.format}
\title{Descriptive "Quantitative" statistics (mean, SD, median...) reporting}
\usage{
apply.format(tibble)
}
\arguments{
\item{tibble}{a tibble object}
}
\value{
A data.frame
}
\description{
\code{apply.format} 

Returns a data.frame with the values of each variable replaced by their corresponding
 formats (label and labels of the tibble).
}
\details{

}
\examples{
 
}
\seealso{
\code{\link{tibble}}
}
