% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at.least.one.R
\name{at.least.one}
\alias{at.least.one}
\title{Create a key from two hierarchical factors}
\usage{
at.least.one(data, subjid = NULL, var, total = FALSE,
  var.label = "var")
}
\arguments{
\item{data}{A data frame}

\item{subjid}{A character. Indicates the ID}

\item{var}{A character. Indicates a factor in the data frame}

\item{total}{Logical. if TRUE the total number of events is calculated}

\item{var.label}{A character. The label of var}
}
\description{
Used to calculate the 'at least one' statistics for several occurence of an event per statistical unit.
}
\details{
Used internally by report.quali.hlev
}
