% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.x.R
\name{check.x}
\alias{check.x}
\title{Check if the variable argument supplied in report.quanti and report.quali is ok}
\usage{
check.x(data, x, substitute = "data")
}
\arguments{
\item{data}{Data.frame object}

\item{x}{Character}

\item{substitute}{Character. The name of the data frame}
}
\description{
Check if the variable argument supplied in report.quanti and report.quali is ok
}
\details{
Used internally by report.quanti and report.quali
}
