% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-data.R
\docType{data}
\name{data}
\alias{data}
\title{Fake clinical data example}
\format{A data frame}
\source{
From pure imagination
}
\usage{
data
}
\description{
Fictive data created only for the purpose of testing the package and showing examples.
}
\examples{
data(data)
\donttest{head(data)}
}
\keyword{datasets}
