% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.key.R
\name{define.key}
\alias{define.key}
\title{Create a key from two hierarchical factors}
\usage{
define.key(data, upper = "", lower = "",
  upper.levels = "Upper.Levels", lower.levels = "Lower.Levels")
}
\arguments{
\item{data}{A data frame}

\item{upper}{A character. Indicates a factor in the data frame which corresponds to the factor with the higher number of levels (typically SOC variable)}

\item{lower}{A character. Indicates a factor in the data frame which corresponds to the factor with the lower number of levels (typically PT variable)}

\item{upper.levels}{A character. The label to be displayed in the table for the upper terms}

\item{lower.levels}{A character. The label to be displayed in the table for the lower terms}
}
\description{
Used to create a single key from two hierarchical factors.
}
\details{
Used internally by report.quali.hlev
}
