% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimnames.desc.R
\name{dimnames.desc}
\alias{dimnames.desc}
\title{dimnames method for desc object}
\usage{
\method{dimnames}{desc}(x)
}
\arguments{
\item{x}{A desc object}
}
\description{
dimnames method for desc object
}
\seealso{
\code{\link{desc}}
}
