% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.desc.R
\name{is.desc}
\alias{is.desc}
\title{Check if it is really a desc object}
\usage{
is.desc(x)
}
\arguments{
\item{x}{Normally, a desc object, but it can be anything...}
}
\description{
Check if it is really a desc object
}
\seealso{
\code{\link{desc}}
}
