% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.label.R
\name{make.label}
\alias{make.label}
\title{Make pretty labels from variable names}
\usage{
make.label(x, l = list(l1 = c("_", " at "), l2 = c(":",
  " interaction with "), l3 = c(".", " ")))
}
\arguments{
\item{x}{Character}

\item{l}{List Each first element is gsubised by the second element of the list}
}
\value{
A character
}
\description{
\code{make.label} 

Mainly used to prettyfy the rownames of anova tables.
}
\details{
Replace with \code{gsub} function each first element of \code{l[[i]]} by the second element of \code{l[[i]]} 
for each element of x.
}
\examples{
data(datafake)
an=anova(lm(y_numeric~y_logistic+GROUP:TIMEPOINT,data=datafake))

# Raw output:
an

rownames(an)=make.label(rownames(an),
l=list(
l1=c("y_logistic","A logistic variable"),
l2=c(":"," interaction with "))
)

# Gives:

an

}
