% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.desc.R
\name{plot.desc}
\alias{plot.desc}
\title{Display the corresponding graphic of statistical table}
\usage{
\method{plot}{desc}(x, ..., title = "", ylim = NULL, xlim = NULL,
  xlab = "", ylab = "", legend.label = "Group", add.sd = F,
  add.ci = F, size.title = 10, add.line = T)
}
\arguments{
\item{x}{A desc object}

\item{...}{Other parameters}

\item{title}{Character The title of the plot}

\item{ylim}{Numeric of length 2 for setting y axis limits}

\item{xlim}{Numeric of length 2 for setting x axis limits}

\item{xlab}{Character Label for x-axis}

\item{ylab}{Character Label for y-axis}

\item{legend.label}{Character Label for the legend (used only if x1 and x2 are not NULL in the desc object)}

\item{add.sd}{Logical. If TRUE it adds bars to the means representing +/-SD (for desc quanti reporting only)}

\item{add.ci}{Logical. If TRUE it adds bars to the means representing 95\% CI (for LS Means reporting only)}

\item{size.title}{Numeric. used to specify the font size of the title}

\item{add.line}{Logical. If TRUE it joins the dots with a line (default to TRUE)}
}
\description{
Display the corresponding graphic of statistical table
}
\details{
It's a wrapper function whiche uses  \code{\link{gg_desc_quanti}} 
 \code{\link{gg_desc_quali}}  or  \code{\link{gg_desc_lsmeans}} depending if
the desc object is of type "quali", "quanti" or "lsmeans"
 (argument\code{type.desc} in \code{\link{desc}} object, see examples below).
}
\examples{
 
data(data)

desc=report.quali(data=data,y="y_logistic",x1="GROUP",
	x2="TIMEPOINT")

plot(desc,
title="Qualitative desc object with 2 explicative variables",
legend.label="Y levels")



desc2=report.quanti(data=data,y="y_numeric",x1="GROUP",
	x2="TIMEPOINT")

plot(desc2,
title="Quantitative desc object with 2 explicative variables",
legend.label="Treatment groups")


#Removing baseline data in the response, for the model

data.mod=droplevels(data[data$TIMEPOINT!="D0",]) 

library(nlme)
library(emmeans)

mod3=lme(y_numeric~baseline+GROUP+TIMEPOINT+GROUP*TIMEPOINT,
random=~1|SUBJID,data=data.mod,na.action=na.omit)

test3=emmeans(mod3,~GROUP|TIMEPOINT)

tab.mod3=report.lsmeans(lsm=test3,at.row="TIMEPOINT")

gg=plot(tab.mod3,title="LS Means plot example")

#Print

gg

}
\seealso{
\code{\link{desc}}
}
