% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyp.R
\name{prettyp}
\alias{prettyp}
\title{Return a p-value vector in nice format}
\usage{
prettyp(p, r = 3)
}
\arguments{
\item{p}{A vector of numerical p-values}

\item{r}{A numeric values that indicates the number of digits to round to}
}
\value{
Return a formated p-value vector with rounded numbers and <0.001 instead of 0.00
}
\description{
\code{prettyp} 
Return a formatted version of pvalues with nice format.
}
\details{
It takes a p and return another one, but better.
}
\examples{

prettyp(c(0.05,0.001,0.00001),3)

}
