% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyround.R
\name{prettyround}
\alias{prettyround}
\title{Return a rounded vector with equal number of digits}
\usage{
prettyround(p, r = 3)
}
\arguments{
\item{p}{Vector of numerical values}

\item{r}{Integer indicates the number of digits to round to}
}
\value{
Return a formatted vector with rounded numbers
}
\description{
\code{prettyround} 
Return a formatted version of the numeric vector.
}
\details{
No detail.
}
\examples{

prettyround(c(-0.05,0.001,0.00001),3)

}
