% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.desc.R
\name{print.desc}
\alias{print.desc}
\title{Print method for desc object}
\usage{
\method{print}{desc}(x, ...)
}
\arguments{
\item{x}{A desc object}

\item{...}{Other parameters}
}
\description{
Print method for desc object
}
\seealso{
\code{\link{desc}}
}
