% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.sessionInfo.R
\name{report.sessionInfo}
\alias{report.sessionInfo}
\title{Export the table of the sessionInfo() to Word or R markdown documents}
\usage{
report.sessionInfo(doc = NULL)
}
\arguments{
\item{doc}{NULL or a rdocx object'}
}
\value{
A flextable object (if doc=NULL) or a rdocx object (if doc= an rdocx object).
}
\description{
\code{report.sessionInfo}

This function enables to export the table of the \code{\link{sessionInfo}}
}
\details{
None
}
\examples{

# For an R markdown document just do
report.sessionInfo()

# Fro a Word document use the doc argument
library(officer)
doc=read_docx()
doc=report.sessionInfo(doc)

}
\seealso{
\code{\link{report.doc}} \code{\link{report.modelinfo}}
}
