% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see.label.R
\name{see.label}
\alias{see.label}
\title{Match between the colnames of the tibble and their corresponding labels.}
\usage{
see.label(tibble)
}
\arguments{
\item{tibble}{A tibble object}
}
\value{
A matrix
}
\description{
\code{see.label} 
enables to see the match between the colnames of the tibble and their corresponding labels.
}
\details{
Usefull function to check the corresponance between colnames and their labels in a tibble object.
}
\examples{
library(haven)

path1 <- system.file("examples", "clinical_sas.sas7bdat", package = "ClinReport")
path2 <- system.file("examples", "formats.sas7bcat", package = "ClinReport")
data=read_sas(path1,path2)

see.label(data)


}
\seealso{
\code{\link{apply.format}}
}
