% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_vline.R
\name{space_vline}
\alias{space_vline}
\title{Compute the indices where to put blanks in report.doc}
\usage{
space_vline(tab, at.row)
}
\arguments{
\item{tab}{data.frame from a desc object}

\item{at.row}{Character. at.row option of a desc object}
}
\value{
A vector of integers
}
\description{
Compute the indices where to put blanks in \code{report.doc}
}
\details{
None
}
\seealso{
\code{\link{report.doc}}
}
