% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_cure-data.R
\docType{data}
\name{time_to_cure}
\alias{time_to_cure}
\title{Fake clinical data example}
\format{A data frame}
\source{
From pure imagination
}
\usage{
time_to_cure
}
\description{
Fictive data created only for the purpose of testing the package and showing examples.
}
\examples{
data(time_to_cure)
\donttest{head(time_to_cure)}
}
\keyword{datasets}
