% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.desc.R
\name{transpose}
\alias{transpose}
\alias{transpose.desc}
\title{Transpose the table of statistics (experimental)}
\usage{
transpose(desc, ...)

\method{transpose}{desc}(desc, ...)
}
\arguments{
\item{desc}{a desc object}

\item{...}{Not used}
}
\value{
A desc object
}
\description{
Used to get the statistics in column for example.

For more examples see the website: \href{https://jfrancoiscollin.github.io/ClinReport}{ClinReport website}
}
\details{
None
}
\examples{

library(reshape2)


data(datafake)

desc=report.quanti(data=datafake,y="y_numeric",x1="GROUP",
	x2="TIMEPOINT",total=TRUE,at.row="TIMEPOINT",subjid="SUBJID")

transpose(desc)

desc=report.quanti(data=datafake,y="y_numeric",x1="GROUP")

transpose(desc)

desc=report.quanti(data=datafake,y="y_numeric")

transpose(desc)

desc=report.quali(data=datafake,y="y_logistic",x1="GROUP")


transpose(desc)


desc=report.quali(data=datafake,y="y_logistic",x1="GROUP",x2="TIMEPOINT")

transpose(desc)



desc=report.quali(data=datafake,y="y_logistic")

transpose(desc)

}
\seealso{
\code{\link{report.quanti}}  \code{\link{report.doc}} \code{\link{desc}}
}
