% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.format.R
\name{apply.format}
\alias{apply.format}
\title{Descriptive "Quantitative" statistics (mean, SD, median...) reporting}
\usage{
apply.format(tibble)
}
\arguments{
\item{tibble}{a tibble object}
}
\value{
A data.frame
}
\description{
\code{apply.format} 

Returns a data.frame with the values of each variable replaced by their corresponding
 formats (label and labels of the tibble).
}
\details{
This function look for the attributes of the tibble object variable and 
apply labels and label to the corresponding variable.
}
\examples{

library(haven)

path1 <- system.file("examples", "clinical_sas.sas7bdat", package = "ClinReport")
path2 <- system.file("examples", "formats.sas7bcat", package = "ClinReport")
data=read_sas(path1,path2)

head(data)
head(apply.format(data))

}
\seealso{
\code{\link{tibble}}
}
