% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Fake clinical data example}
\format{A data frame
\describe{
  \item{y_numeric}{Fake numerical response}
  \item{y_logistic}{Fake Logistic response}
  \item{y_poisson}{Fake Poisson response}
  \item{baseline}{A baseline covariate}
  \item{VAR}{A factor}
  \item{GROUP}{A fake treatment factor}
  \item{TIMEPOINT}{A fake time factor}
  \item{SUBJID}{A fake id factor}
}}
\usage{
data
}
\description{
Fictive data created only for the purpose of testing the package and showing examples.
}
\keyword{datasets}
