% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc}
\alias{desc}
\title{Constructor function for the desc object}
\usage{
desc(output = NULL, total = NULL, nbcol = NULL, y = NULL,
  x1 = NULL, x2 = NULL, at.row = NULL, subjid = NULL,
  type.desc = NULL, type = NULL, y.label = NULL, type.mod = NULL,
  raw.output = NULL, contrast = NULL, contrast.name = NULL,
  stat.name = NULL)
}
\arguments{
\item{output}{a data.frame object}

\item{total}{Boolean. Indicates if a column Total should be added or not}

\item{nbcol}{Numeric. number of columns that are not statistics}

\item{y}{Character indicating the response variable}

\item{x1}{Character indicating a factor in the data}

\item{x2}{Character indicating a factor in the data}

\item{at.row}{Character. Pass to spacetable function. Used to space the results (see \code{\link{spacetable}})}

\item{subjid}{Character. Indicates in the data.frame the name of the column used to identify the Id of the subjects.
If not null, it adds in the headers the number of unique subject per levels of x1 or y (if x1 is null).}

\item{type.desc}{Character. Can be "quali" "quanti" "lsmeans" or "quali_quanti"}

\item{type}{Character. Can be "response" or "link"}

\item{y.label}{Character. The label for y response variable}

\item{type.mod}{Character. The type of models the LS means are coming from (can be 'quali' or 'quanti')}

\item{raw.output}{Data.frame. The raw statistics unformatted from \code{report.quali}, \code{report.quanti} and  \code{report.lsmeans}}

\item{contrast}{Logical. Specify if the contrast function has been used after the emmeans function}

\item{contrast.name}{Character. Corresponds to the label of the column in which the contrasts are specified
Default value is 'contrast'.}

\item{stat.name}{Character. Indicates the name of the variable that report the statistics in report.quanti function.
Default = "Statistics"}
}
\value{
A desc object
}
\description{
Creates a desc object.
}
\details{
Used internally by code{report.XXX} functions
}
\examples{

\dontrun{

x=desc()

}

}
\seealso{
\code{\link{report.quali}} \code{\link{report.quanti}} \code{\link{report.lsmeans}}
}
