% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see.label.R
\name{see.label}
\alias{see.label}
\title{Match between the colnames of the tibble and their corresponding labels.}
\usage{
see.label(tibble)
}
\arguments{
\item{tibble}{A tibble object}
}
\value{
A matrix
}
\description{
\code{see.label} 
enables to see the match between the colnames of the tibble and their corresponding labels.
}
\details{
No details
}
\examples{
# Not yet


}
\seealso{
\code{\link{apply.format}}
}
