% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_to_cure}
\alias{time_to_cure}
\title{Fake survival data set}
\format{A data frame}
\usage{
time_to_cure
}
\description{
Simulated survival data created only for the purpose of testing the package and showing examples.
}
\keyword{datasets}
