% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_to_cure}
\alias{time_to_cure}
\title{time_to_cure}
\format{A data frame
#' \describe{
 \item{time}{Fake numerical time values}
\item{status}{Fake status values, 1=cured 0=not cured}
\item{block}{Fake factor}
\item{random_number}{Fake factor}
\item{Sex}{Fake sex factor}
\item{Weight_D0}{Fake weight covariate}
\item{Pen}{Fake factor}
\item{TCS}{Fake covariate}
\item{Group}{Fake treatment group factor}
\item{Subjid}{Fake subject Id}
}}
\usage{
data(time_to_cure)
}
\description{
Simulated survival data created only for the purpose of testing the package and showing examples.
}
\keyword{datasets}
